Разбивает файл js на модули, добавляет точку входа
